+++ src/pages/api/category/delete-user-category.ts
@@ -92,7 +92,7 @@ export default async function handler(
 		await supabase
 			.from(SHARED_CATEGORIES_TABLE_NAME)
 			.delete()
-			.match({ category_id: request.body.category_id });
+			.match({ category_id: request.body.category_id, user_id: userId });
 
 	if (!isNull(sharedCategoryError)) {
 		response.status(500).json({
@@ -146,6 +146,7 @@ export default async function handler(
 		.from(CATEGORIES_TABLE_NAME)
 		.delete()
 		.eq("id", request.body.category_id)
+		.eq("user_id", userId)
 		.select(`*`);
 
 	if (isNull(data)) {
